refactor some components as sections as stand-alone files
